backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
SQLAlchemy==2.0.23
chromadb==0.4.18
openai==1.3.7  # 火山引擎兼容 OpenAI SDK
tiktoken==0.5.2
python-multipart==0.0.6
